Added additional functionality to remove sprint method.

+++ MyApplication/app/src/main/java/com/actual/myapplication/IdGenerator.java
@@ -19,7 +19,7 @@ package com.actual.myapplication;
         IdGenerator.strategy = strategy;
     }
 
-    public static long generateId() {
+    public static int generateId() {
         return strategy.generateId();
     }
